Add files via upload
Exploratory Data Analysis Lab
Estimated time needed: 30 minutes

In this module you get to work with the cleaned dataset from the previous module.

In this assignment you will perform the task of exploratory data analysis. You will find out the distribution of data, presence of outliers and also determine the correlation between different columns in the dataset.

Objectives
In this lab you will perform the following:

Identify the distribution of data in the dataset.

Identify outliers in the dataset.

Remove outliers from the dataset.

Identify correlation between features in the dataset.