Update README.md
In this module, you will create a dashboard using IBM Cognos Analytics. This platform will give you the ability to create various charts while assembling a dashboard that is appealing and easy to understand. Your dashboard will contain your data analysis, which should be intuitive and allow for the drill-down of data.
Learning Objectives
Utilize the features in IBM Cognos to improve your charts.
Demonstrate the use of correct chart types to use to visualize data.
Create a dashboard that helps visualize the key points of the survey dataset.